# web/src/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Realtor Mogul API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/buy-boxes": {
      "get": {
        "tags": [
          "screener"
        ],
        "summary": "List Boxes",
        "operationId": "list_boxes_buy_boxes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BuyBoxOut"
                  },
                  "type": "array",
                  "title": "Response List Boxes Buy Boxes Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "screener"
        ],
        "summary": "Create Box",
        "operationId": "create_box_buy_boxes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyBoxIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuyBoxOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/buy-boxes/{box_id}": {
      "put": {
        "tags": [
          "screener"
        ],
        "summary": "Update Box",
        "operationId": "update_box_buy_boxes__box_id__put",
        "parameters": [
          {
            "name": "box_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Box Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyBoxIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuyBoxOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "screener"
        ],
        "summary": "Delete Box",
        "operationId": "delete_box_buy_boxes__box_id__delete",
        "parameters": [
          {
            "name": "box_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Box Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/buy-boxes/{box_id}/recommendations": {
      "get": {
        "tags": [
          "screener"
        ],
        "summary": "Recommendations",
        "operationId": "recommendations_buy_boxes__box_id__recommendations_get",
        "parameters": [
          {
            "name": "box_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Box Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/buy-boxes/{box_id}/seen": {
      "post": {
        "tags": [
          "screener"
        ],
        "summary": "Mark Seen",
        "operationId": "mark_seen_buy_boxes__box_id__seen_post",
        "parameters": [
          {
            "name": "box_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Box Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/listings/import": {
      "post": {
        "tags": [
          "screener"
        ],
        "summary": "Import Listings",
        "description": "Import a listings CSV (e.g. a Redfin \"Download All\" export of your search).",
        "operationId": "import_listings_listings_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingImportIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingImportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/listings/{listing_id}": {
      "get": {
        "tags": [
          "screener"
        ],
        "summary": "Listing Detail",
        "operationId": "listing_detail_listings__listing_id__get",
        "parameters": [
          {
            "name": "listing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Listing Id"
            }
          },
          {
            "name": "buy_box_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Buy Box Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "screener"
        ],
        "summary": "Patch Listing",
        "description": "Override the rent estimate or the unit count. Send rent_override=null to clear.",
        "operationId": "patch_listing_listings__listing_id__patch",
        "parameters": [
          {
            "name": "listing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Listing Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/listings/{listing_id}/watchlist": {
      "post": {
        "tags": [
          "screener"
        ],
        "summary": "Add To Watchlist",
        "description": "Save the listing, underwritten with the buy box's assumptions, as a watchlist deal.",
        "operationId": "add_to_watchlist_listings__listing_id__watchlist_post",
        "parameters": [
          {
            "name": "listing_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Listing Id"
            }
          },
          {
            "name": "buy_box_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Buy Box Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Add To Watchlist Listings  Listing Id  Watchlist Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BuyBoxIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "criteria": {
            "$ref": "#/components/schemas/Criteria"
          },
          "assumptions": {
            "$ref": "#/components/schemas/Assumptions"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "BuyBoxIn"
      },
      "BuyBoxOut": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "criteria": {
            "$ref": "#/components/schemas/Criteria"
          },
          "assumptions": {
            "$ref": "#/components/schemas/Assumptions"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "last_viewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Viewed At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "last_viewed_at"
        ],
        "title": "BuyBoxOut"
      },
...
      "Criteria": {
        "properties": {
          "market_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Market Ids",
            "description": "Empty = any market"
          },
          "property_types": {
            "items": {
              "type": "string",
              "enum": [
                "single_family",
                "multi_2_4",
                "condo",
                "multifamily",
                "commercial",
                "land"
              ]
            },
            "type": "array",
            "title": "Property Types"
          },
          "min_price": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Price"
          },
          "max_price": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Price",
            "default": 500000
          },
          "min_beds": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Beds"
          },
          "max_days_on_market": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Days On Market"
          },
          "min_year_built": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Year Built"
          },
          "min_levered_irr": {
            "type": "number",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Min Levered Irr",
            "default": 0.12
          },
          "min_cash_on_cash": {
            "type": "number",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Min Cash On Cash",
            "default": 0.06
          },
          "min_dscr": {
            "type": "number",
            "maximum": 5.0,
            "minimum": 0.0,
            "title": "Min Dscr",
            "default": 1.2
          },
          "min_cap_rate": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Cap Rate"
          }
        },
        "type": "object",
        "title": "Criteria"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ListingDetail": {
        "properties": {
          "listing": {
            "$ref": "#/components/schemas/ListingOut"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/ListingEventOut"
            },
            "type": "array",
            "title": "Events"
          },
          "rent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RentEstimate"
              },
              {
                "type": "null"
              }
            ]
          },
          "evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Evaluation"
              },
              {
                "type": "null"
              }
            ]
          },
          "filtered_because": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Filtered Because"
          }
        },
        "type": "object",
        "required": [
          "listing",
          "events",
          "rent",
          "evaluation",
          "filtered_because"
        ],
        "title": "ListingDetail"
      },
...
      "ListingImportIn": {
        "properties": {
          "csv": {
            "type": "string",
            "maxLength": 20000000,
            "minLength": 1,
            "title": "Csv"
          }
        },
        "type": "object",
        "required": [
          "csv"
        ],
        "title": "ListingImportIn"
      },
      "ListingImportOut": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "new": {
            "type": "integer",
            "title": "New"
          },
          "updated": {
            "type": "integer",
            "title": "Updated"
          },
          "price_changes": {
            "type": "integer",
            "title": "Price Changes"
          }
        },
        "type": "object",
        "required": [
          "source",
          "new",
          "updated",
          "price_changes"
        ],
        "title": "ListingImportOut"
      },
      "ListingOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "zip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zip"
          },
          "market_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Id"
          },
          "market_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Name"
          },
          "property_type": {
            "type": "string",
            "title": "Property Type"
          },
          "units": {
            "type": "integer",
            "title": "Units"
          },
          "units_inferred": {
            "type": "boolean",
            "title": "Units Inferred"
          },
          "beds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beds"
          },
          "baths": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baths"
          },
          "sqft": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sqft"
          },
          "year_built": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year Built"
          },
          "hoa_monthly": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hoa Monthly"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "days_on_market": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days On Market"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "stated_rent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stated Rent"
          },
          "rent_override": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rent Override"
          },
          "price_change": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Change"
          },
          "first_seen": {
            "type": "string",
            "format": "date-time",
            "title": "First Seen"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "address",
          "city",
          "state",
          "zip",
          "market_id",
          "market_name",
          "property_type",
          "units",
          "units_inferred",
          "beds",
          "baths",
          "sqft",
          "year_built",
          "hoa_monthly",
          "price",
          "status",
          "days_on_market",
          "url",
          "stated_rent",
          "rent_override",
          "price_change",
          "first_seen"
        ],
        "title": "ListingOut"
      },
      "ListingPatch": {
        "properties": {
          "rent_override": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rent Override"
          },
          "units": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Units"
          }
        },
        "type": "object",
        "title": "ListingPatch"
      },
...
      "RecommendationsOut": {
        "properties": {
          "buy_box": {
            "$ref": "#/components/schemas/BuyBoxOut"
          },
          "summary": {
            "$ref": "#/components/schemas/Summary"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/RecommendationOut"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "buy_box",
          "summary",
          "rows"
        ],
        "title": "RecommendationsOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}